add only author can delete question scenario

+++ app/controllers/questions_controller.rb
@@ -1,9 +1,9 @@
 class QuestionsController < ApplicationController
-  before_action :authenticate_user!, only: %i[new create]
+  before_action :authenticate_user!, only: %i[new create destroy]
 
   expose :questions, -> { Question.all }
   expose :question,
-        build: ->(params, _scope) { current_user.questions.new(params) }
+         build: ->(params, _scope) { current_user.questions.new(params) }
   expose :answers, -> { question.answers }
   expose :answer, -> { question.answers.new }
 
@@ -24,7 +24,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user.author_of?(question)
+      question.destroy
+      flash[:notice] = 'Question delete successfully'
+    else
+      flash[:alert] = "You can not delete someone else's question"
+    end
     redirect_to questions_path
   end
